requirements.txt
Django
requests
python_dotenv
mysqlclient
beautifulsoup4
discord